requirements.txt
fastapi
uvicorn
pydantic    
httpx
beautifulsoup4
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
tweepy
aiofiles
pytest
jinja2
python-multipart
gunicorn
prometheus-client